fix: Improve .svelte file resolution in the Vite plugin (#28)
This PR makes the .svelte file resolution more robust.

While testing with a SvelteKit project, I discovered the current
implementation of the Vite plugin doesn’t work when generating the
client-side code. The issue arises from an inconsistency in the paths
provided to the `resolveId` method:

- For server-side code generation, both `source` and `importer` are
absolute paths:
     `source = "/User/path/to/repo/src/lib/components/Component.svelte"`

- For client-side code generation, `source` becomes an _absolute_ path
starting from the project root:
      `source = "/src/lib/components/Component.svelte"`