test: Improve snapshots serializer - `<semver>` (#30)
In PR #19 I noticed that we'll be bothered to update snapshots everytime
we update dependencies.

See the comment with example:
<https://github.com/svelte-docgen/svelte-docgen/pull/19#discussion_r1885550919>

So, this gives extra ability to the current snapshots serializer to wrap
any semver pattern with `<semver>`.
Information about version of either `svelte`, `typescript` or anything
else is more important on the user side.